+++ skill_sdk/log.py
@@ -10,6 +10,7 @@
 """Logging"""
 
 import os
+import re
 import time
 import json
 import logging.config
@@ -30,6 +31,7 @@ def setup_logging(
     :param log_format:
     :return:
     """
+    from uvicorn.config import LOGGING_CONFIG
 
     log_level = log_level or config.settings.LOG_LEVEL
     log_format = log_format or config.settings.LOG_FORMAT
@@ -39,6 +41,18 @@ def setup_logging(
 
     try:
         logging.config.dictConfig(get_config_dict(log_level, log_format))
+
+        # Patch Uvicorn logger default formats and level
+        LOGGING_CONFIG["formatters"]["access"]["()"] = (
+            "skill_sdk.log.CloudGELFFormatter"
+            if log_format == config.FormatType.GELF
+            else "uvicorn.logging.AccessFormatter"
+        )
+
+        LOGGING_CONFIG["loggers"]["uvicorn"]["level"] = log_level
+        LOGGING_CONFIG["loggers"]["uvicorn.error"]["level"] = log_level
+        LOGGING_CONFIG["loggers"]["uvicorn.access"]["level"] = log_level
+
     except KeyError:
         raise RuntimeError("Invalid log format: %s", repr(log_format))
 
@@ -174,11 +188,40 @@ patch_logger()
 
 ###############################################################################
 #                                                                             #
-#  Limit log message size                                                     #
+#   Optional formatter for GunicornLogger                                     #
 #                                                                             #
 ###############################################################################
+try:
+    from gunicorn.glogging import Logger
+
+    class GunicornLogger(Logger):
+        """Gunicorn logger that formats log messages with CloudGELFFormatter"""
+
+        def setup(self, cfg):
+            self.loglevel = config.settings.LOG_LEVEL
+            self.error_log.setLevel(self.loglevel)
+            self.access_log.setLevel(self.loglevel)
+
+            if config.settings.LOG_FORMAT == config.FormatType.GELF:
+                self._set_handler(self.error_log, cfg.errorlog, CloudGELFFormatter())
+                self._set_handler(self.access_log, cfg.errorlog, CloudGELFFormatter())
+
+
+except ModuleNotFoundError:  # pragma: no cover
+    pass
+
+
+###############################################################################
+#                                                                             #
+#  Limit log message size, and hide tokens from logs                          #
+#                                                                             #
+###############################################################################
+JWT_REGEX = re.compile(r"^eyJ[A-Za-z0-9_-]*\.[A-Za-z0-9._-]*")
+
+
 def _trim(s):
     """Trim long string to LOG_ENTRY_MAX_STRING(+3) length"""
+
     return (
         s
         if not isinstance(s, str) or len(s) < config.settings.LOG_ENTRY_MAX_STRING
@@ -186,22 +229,25 @@ def _trim(s):
     )
 
 
-def _copy(d):
-    """Recursively copy dictionary values, trimming long strings"""
+def _copy(d, hide_tokens: bool = False):
+    """
+    Recursively copy dictionary values, trimming long strings, and hiding possible JWT tokens
+    """
 
     if isinstance(d, dict):
         return {k: _copy(v) for k, v in d.items()}
     elif isinstance(d, (list, tuple)):
         return [_copy(v) for v in d]
     else:
-        return _trim(d)
+        return _trim(JWT_REGEX.sub("eyJ*****", d) if hide_tokens else d)
 
 
-def prepare_for_logging(record):
+def prepare_for_logging(record, hide_tokens: bool = False):
     """
     Trim long strings before logging a record
 
-    :param record:  value to log
+    :param record:      value to log
+    :param hide_tokens: specifies if JWT-like values should be replaced by asterisks
     :return:
     """
-    return _copy(record)
+    return _copy(record, hide_tokens)