+++ CHANGELOG.md
@@ -4,7 +4,43 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+
+### Bugfixes
+
+- **Logging:**
+
+    - Fix uvicorn access log formatter to consistently use either human-readable or GELF format.
+
+    - Add logging helpers to possibly hide JWT-like token strings.
+
+
+## 1.0.4 - 2021-05-26
+
+### Bugfixes
+
+- Fix the initialization of Prometheus metrics exporter endpoint.
+
+
+- **Logging:**
+
+    - Display uvicorn access log in GELF format.
+      
+    - Add Gunicorn logging formatter: can be used to export skill logs as GELF-compatible JSON when deploying with Gunicorn.  
+    To activate, add [`--logger-class=skill_sdk.log.GunicornLogger`](https://docs.gunicorn.org/en/stable/settings.html#logger-class) parameter when deploying the skill.
+  
+
+## 1.0.3 - 2021-05-25
+
+### Features
+
+
+- Skill Designer UI: a tool for rapid skill prototyping.
+    Start a skill with `vs develop [impl]` and hit [http://localhost:4242](http://localhost:4242) 
+    to access the UI from a browser.
+  
+
 - Support local skill translations in [Rails-compatible](https://guides.rubyonrails.org/i18n.html) YAML format.
+
 
 ## 1.0.2 - 2021-05-05
 
@@ -22,7 +58,7 @@ All notable changes to this project will be documented in this file.
     - `LocationService.device_location` retrieves the device location (the info, a user has set up in companion app) with geo coordinates.
 
     - `LocationService.address_lookup` returns a list of addresses (with geo-coordinates) for a given query.
-    A query consists of any of address fields (country, zip, street name, house number).
+    A query consists the address fields (country, zip, street name, house number).
 
 ### Bugfixes